py/retake/sage/requirements.txt
boto3==1.28.12
botocore==1.31.12
dataclasses-json
fastapi
ffmpeg-python
modal-client
numpy
openai
opencv-python
pandas==2.0.3
python-dotenv==1.0.0
uvicorn